--- cadquery/occ_impl/types.py
@@ -1,3 +1,0 @@
-from typing_extensions import Literal
-
-UnitLiterals = Literal["MM", "CM", "M", "KM", "INCH", "FT", "MI", "UM", "NM"]
