+++ cadquery/types.py
@@ -1,3 +1,4 @@
-from typing import Union
+from typing import Union, Literal
 
+UnitLiterals = Literal["MM", "CM", "M", "KM", "INCH", "FT", "MI", "UM", "NM"]
 Real = Union[int, float]
